Change how flags are handled
Flags are now passed in as options to the script, and handled by default if
they are not present.

Supported flags are -r|--use-runtime -t|--close-to-tray -f|--fx -h|--help

The --fix flag displays fixes for common issues experienced in games on Steam
on Linux.

So far, Borderlands 2 is supported in the sense that it had an issue that
was able to be corrected with a non-intuitive work around.

The --help flag displays all currently supported options